requirements.txt
langchain>=0.1.0
langgraph>=0.0.40
langchain_core>=0.1.0
openai>=1.0.0
python-dotenv>=1.0.0
selenium>=4.15.0
flask>=3.0.0
pytest>=7.4.0
requests>=2.31.0
pyyaml>=6.0.1
streamlit>=1.28.0
sqlalchemy>=2.0.0
beautifulsoup4>=4.12.0
langchain-community>=0.0.10
tenacity>=8.2.0
apscheduler>=3.10.0
# Google Sheets API için bağımlılıklar
gspread>=5.10.0
google-auth>=2.22.0
google-api-python-client>=2.100.0
# Scraper için ek bağımlılıklar
webdriver-manager>=4.0.0
lxml>=4.9.0
cloudscraper>=1.2.71
fake-useragent>=1.3.0
anticaptchaofficial>=1.0.55 # CAPTCHA çözümü için
2captcha-python>=1.1.3 # Alternatif CAPTCHA çözüm servisi
pandas>=2.0.0 # Veri işleme ve düzenleme için
